PlatformIO/Projects/Spirob_aktuation/pyproject.toml
[project]
name = "spirob-aktuation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
    "customtkinter>=5.0",
    "polars>=1.38.1",
    "matplotlib>=3.10.8",
    "scipy>=1.15.3",
    "pyqt6>=6.10.2",
    "pyqtgraph>=0.14.0",
    "mujoco>=3.5.0",
]
